First Generic Naive Callbacks
Generic programming is pretty cool once you start to grasp the patterns
and paradigms.  First version of naiveMatch.  Todo: don't use a struct
with Match.  Instead make it into a class and create a class hierarchy
with subtypes providing more information about the match - useful for
the coming more advanced string matching algorithms that search on
multiple texts or patterns or ones that search for substring or even
approximate matches.